requirements.txt
# Simulation
jax==0.4.23
jaxlib==0.4.23
jax-md==0.2.8
scipy==1.12.0
flax

# Interface
panel==1.3.8
param==2.0.2
matplotlib==3.8.2

# Client-server communication and subprocesses
grpcio==1.60.0
grpcio-tools
psutil

# Tests and code formatting
pytest==8.0.0
black

# Configurations
hydra-core

# Notebooks
notebook